job_application_assistant/requirements.txt
gradio==5.47.2
langchain==0.3.27
langchain-community==0.3.30
google-search-results==2.4.2
langgraph==0.6.6
python-docx==1.2.0